user auth functioning somewhat
