+++ public/stylesheets/style.css
@@ -8,10 +8,11 @@
 
 body {
   background-color: #772014;
-  color: rgb(75, 33, 0);
+  color: white;
   font-family: 'Helvetica', 'Arial', sans-serif;
   box-sizing: border-box;
   width: 100%;
+
 }
 
 #page {
@@ -93,6 +94,7 @@ body {
   border-radius: .5rem;
   padding: 1%;
   background-color: white;
+  color: #772014;
   text-align: left;
 }
