+++ views/sign-in.hbs
@@ -1,5 +1,5 @@
 <form id="sign-in-form" action="/sign-in" method="POST">
   <input type="text" name="email" placeholder="Email" required>
-  <input type="text" name="password" placeholder="Password" required>
+  <input type="password" name="password" placeholder="Password" required>
   <button type="submit">Log In</button>
 </form>
