+++ views/sign-up.hbs
@@ -1,6 +1,6 @@
 <form id="sign-up-form" method="POST" action="/sign-up">
   <input type="text" name="email" placeholder="Your Email">
-  <input type="text" name="password" placeholder="New Password">
+  <input type="password" name="password" placeholder="New Password">
   <input type="text" name="screenName" placeholder="Screen Name">
   <button type="submit">Create Profile</button>
 </form>
